Remove previous experiments
