--- rzk/src/Rzk/Simple/Syntax/Term/Example.hs
@@ -1,5 +1,0 @@
-module Rzk.Simple.Syntax.Term.Example where
-
-import           Rzk.Simple.Syntax.Term
-
-
